Group Project
This project was created in a team of 4 individuals named Jason, Henry,
Jay, and Johnathan. The project itself is a ecommerce website called J3H
that allows users to purchase electronics, movies and games. Tools that
were used to complete the project are: HTML5/CSS3, JSP, Java, MySQL,
Photoshop, JScript, and used Eclipse IDE.